Add support for optional node data
Add optional data parameter for the addNode method. If data is not given
the node name is used as data.

Add methods setNodeData and getNodeData to manipulate the data
associated with a node.

Update the hasNode method to be able to cope with falsy node data.

Update README to document the new functionality.

Fixes #12