Allows for circular dependencies
When working with ES Module dependencies, this graph is very handy. I'd
like to expand it to support circular dependencies, which aren't an
issue in this realm.

I don't think this API is the most ideal, since it tacks on yet another
argument to all the calls. I didn't add tests yet, since I wanted to
gauge your opinion.